clean up sdf and adapter
